moved files from Extraction folder to src, split extraction package to extration and github
