+++ src/extraction_old/Extraction.java
@@ -1,0 +1,38 @@
+package extraction;
+import github.Repository;
+import java.util.ArrayList;
+
+
+/**
+ * Extraction class.
+ * @author Thomas Neumayr
+ * @version 1.0
+ */
+public class Extraction
+{
+
+	private ArrayList<Repository> repositories;
+	/**
+	* Constructor for extraction
+	* Sets repo arraylist
+	*/
+	public Extraction()
+	{
+		repositories = new ArrayList<>();
+	}
+
+
+	public void addRepo(Repository repo)
+	{
+		repositories.add(repo);
+	}
+
+    /**
+    * Function to print the string "Hello, world!" to the console screen.
+    * @param args Unused arguments.
+    */
+    public static void main(String[] args) 
+    {
+        System.out.println("Hello, world!");
+    }
+}
