+++ src/extraction_old/ExtractionTest.java
@@ -1,9 +1,8 @@
-package extraction;
 /**
  * ExtractionTest
  *
  */
-
+package extraction;
 // Stuff to redirect System.out for testing purposes.
 import java.lang.IllegalArgumentException;
 import java.io.PrintStream;
@@ -18,9 +17,8 @@ import static org.junit.Assert.*;
 public class ExtractionTest {
 
     @Test
-    public void testExtraction() {
-    
-    /*
+    public void testPrint() {
+
 	// Prep for test
 	// Actual and expected outputs
 	String correctResult = "Hello, world!\n";
@@ -78,8 +76,6 @@ public class ExtractionTest {
 	    System.err.println("Result: PASSED\n");
 
 	assertEquals(testFailed, null);
-    */
-    
     }
