+++ src/github/IssueTest.java
@@ -1,4 +1,4 @@
-package extraction;
+package github;
 /**
  * Issue Test
  *
